+++ src/strongbox/wowinterface_api.clj
@@ -28,25 +28,23 @@
     (error (utils/reportable-error "given addon-summary has no game track list."))
     (error addon-summary))
 
-  ;; todo: this shouldn't be an 'if' if there is no 'else'
+  ;; when the selected game track is supported by the addon, check for updates.
   (when (some #{game-track} (:game-track-list addon-summary))
     (let [url (str wowinterface-api "/filedetails/" (:source-id addon-summary) ".json")
-          result-list (some-> url http/download-with-backoff http/sink-error utils/from-json)
-          result (first result-list)]
-      (when result
+          result-list (some-> url http/download-with-backoff http/sink-error utils/from-json)]
+      (when-not (empty? result-list)
 
-        ;; has this happened before? can we find an example?
-        (when (> (count result-list) 1)
-          (warn "wowinterface api returned more than one result for addon with id:" (:source-id addon-summary)))
-
-        (let [sid (:source-id addon-summary)
-              ;; rarely present. use it if found. actual value of `aid` not necessary, it seems to work when empty as well.
-              aid (extract-aid (:UIDownload result))]
-          [{:download-url (if aid
-                            (format "https://cdn.wowinterface.com/downloads/getfile.php?id=%s&aid=%s" sid aid)
-                            (format "https://cdn.wowinterface.com/downloads/getfile.php?id=%s" sid))
-            :version (:UIVersion result)
-            :game-track game-track}])))))
+        ;; 2023-06-09: we don't expect more than one result from wowi, ever, but for the sake of testing and
+        ;; consistency with other hosts it is now supported.
+        (mapv (fn [result]
+                (let [sid (:source-id addon-summary)
+                      ;; rarely present. use it if found. actual value of `aid` not necessary, it seems to work when empty as well.
+                      aid (extract-aid (:UIDownload result))]
+                  {:download-url (if aid
+                                   (format "https://cdn.wowinterface.com/downloads/getfile.php?id=%s&aid=%s" sid aid)
+                                   (format "https://cdn.wowinterface.com/downloads/getfile.php?id=%s" sid))
+                   :version (:UIVersion result)
+                   :game-track game-track})) result-list)))))
 
 (defn-spec parse-user-string (s/or :ok :addon/source-id :error nil?)
   "extracts the addon ID from the given `url`"